// src/app/models/admin/capphephoatdongkhoangsan/cpkhaithackhoangsan/cpkhaithackhoangsan.model.ts
export class OutputCpKhaiThacKhoangSanModel {
    idcapphepkhaithac: string;
    idmo: string;
    sohieu: string;
    tenmo: string;
    diadiem: string;
    duancongtrinh: string;
    dientichkhaithac: number;
    dientichtralai: number;
    truluongdiachat: number;
    truluongkhaithac: number;
    truluongtralai: number;
    thoihankhaithac: number;
    ngaybdkhaithac: string;
    ngayktkhaithac: string;
    phuongphapkhaithac: number;
    congsuatkhaithac: number;
    mucsaukhaithactu: number;
    mucsaukhaithacden: number;
    dangkhoangsan: number;
    donvitruluong: string;
    donvicongsuat: string;
    donvidientich: string;
    donvithoihan: string;
    donvichieusau: string;
    geom: string;
    geovn: string;
    geowgs: string;
    hequychieu: string;
    idgiayphep: string;
    siteid: string
}

export class InputCpKhaiThacKhoangSanModel {
    idcapphepkhaithac: string;
    tenmo: string;
    sohieu: string;
    diadiem: string;
    duancongtrinh: string;
    dientichkhaithac: number;
    dientichtralai: number;
    truluongdiachat: number;
    truluongkhaithac: number;
    truluongtralai: number;
    thoihankhaithac: number;
    ngaybdkhaithac: string;
    ngayktkhaithac: string;
    phuongphapkhaithac: number;
    congsuatkhaithac: number;
    mucsaukhaithactu: number;
    mucsaukhaithacden: number;
    dangkhoangsan: number;
    donvitruluong: string;
    donvicongsuat: string;
    donvidientich: string;
    donvithoihan: string;
    donvichieusau: string;
    hequychieu: string;
    idgiayphep: string;
}